Crate fp_newton

Arithmetic of formal power series.

Data representation

In fact a formal power series $f = f_0 x + f_1 x + f_2 x^2 + \dots \in \mathbb{F}_p[[x]]$ has an infinite number of coefficients, but it is too hard for computers to handle it. Therefore, we represent a formal power series $f$ as a finite sequence $f_0, f_1, \dots, f_{d-1}$ of length $d$. This $d$ is called precision. Trailing zeros are also allowed in this representation.

Furthermore, higher-degree coefficients are regarded as zero.

In conclution, the set of values that Vec<Fp<P>> can represent is exactly the same as the following set:

$$ \bigsqcup_{d=0}^\infty \mathbb{F}_p[x] / (x^d) $$

Especially, any two intances of different precisions are assumed not to be equal to each other.

Note on complexity

We only consider cases where the precision is a power of 2. If the precision is not a power of 2, the complexity is twice as bad as otherwise.

We may assume the complexity $\mathcal{M}(d)$ of multiplication of two polynomials of degree $d$ and the complexity $\mathcal{F}(d)$ of FFT/IFFT of length $d$ satisfies the following property:

$$ \begin{aligned} \mathcal{M}(d) &= \Theta(d\log d) , \mathcal{F}(d) = \Theta(d\log d), \\ \mathcal{M}(d) &= 3 \mathcal{F}(2d) + O(d) \ \left( = 6 \mathcal{F}(d) + O(d) \right). \end{aligned} $$

If each step in Newton’s method is performed in $a\mathcal{M}(d)$ time (where $d$ is the resulting precision of each iteration), then the total complexity is $2a\mathcal{M}(d) + O(d)$.

We omit $O(d)$ terms when it is not important.

$O(d)$ is omitted here.

Name	Complexity

fps_inv	$2\mathcal{M}(d)$

fps_sqrt	$6\mathcal{M}(d)$

fps_deriv	$O(d)$

fps_int	$O(d)$

fps_log	$3\mathcal{M}(d)$

fps_exp	$(10+2/3)\mathcal{M}(d)$

Macros

fps

Define a formal power series in the same way as vec!.

Functions

fps_deriv

Returns the derivative of a formal power series.

fps_exp

Returns the exponential of a formal power series.

fps_int

Returns the integral of a formal power series with zero constant term.

fps_inv

Returns the multiplicative inverse of a formal power series.

fps_log

Returns the logarithm of a formal power series.

fps_pow

Resutns the $n$-th power of a formal power series.

fps_sqrt

Returns the square root of a formal power series.